+++ example/ionic-angular-v7/src/app/app.module.ts
@@ -29,6 +29,12 @@ Sentry.init(
         blockAllMedia: true,
       }),
     ],
+    _experiments: {
+      enableLogs: true,
+      beforeSendLog: (log) => {
+        return log;
+      }
+    },
     // A release identifier
     release: '1.0.0',
     // A dist identifier